Block counter field not required as part of AES-CTR IV
Previously, the library required application to provide 16 byte IV with the following layout:
- Nonce (4 bytes)
- ESP IV (8 bytes)
- block counter (4 bytes)

After the change, an application only needs to provide 12 byte IV consisting of:
- Nonce (4 bytes)
- ESP IV (8 bytes)

The block counter filed is initialized to 0x00000001 with the library.
Consequently, library reads only 12 bytes at the location pointed by an application.

Applications that provide pointer to a 16 byte IV will still work without modification
as long as the block counter was set to 0x00000001 by them.