Updated LibPerfApp to allocate MB manager memory on main thread.
This is required to fix a segmentation fault caused by the following command:
$ ./ipsec_perf --threads 8

This happens only on some platforms and only for number of threads equal to 8.
The issue is believed to be caused by the libc memory allocator incorrectly
allocating memory under certain corner conditions.
The workaround is to allocate MB manager memory sequentially in the main thread.

Change-Id: I3c5f6efb9372d67c0d80eeed06c14a6a74ad2baa