Optimize AVX512 16x16 transpose
SHA1, SHA224 and SHA256 implementations perform
a 16x16 32-bit transpose of the data coming from 16 lanes,
before starting to process it.

This transpose is optimized by loading the first half of the 64-byte blocks
in all the lanes in the 8 ZMM registers and the other second half
in other 8 ZMM registers, avoiding a last round of shuffles required
to transpose all the data.

Change-Id: I603273f346418e5be0cd57f84a4f19c78084015c