Optimize AVX512 8x8 transpose
SHA384 and SHA512 implementations perform an 8x8 64-bit transpose of the data
coming from 8 lanes, before starting to process it.

This transpose is optimized by loading the first half of the 64-byte blocks
in all the lanes in the 4 ZMM registers and the other second half
in other 4 ZMM registers, avoiding a last round of shuffles required
to transpose all the data.

Change-Id: Iafdb076f9d5a6a3a95e0a0221543ef3397aa804b