Optimize AVX2 8x8 transpose
SHA1, SHA256 and MD5 implementations perform an 8x8 32-bit transpose
of the data coming from 8 lanes, before starting to process it.

This transpose is optimized by loading the first half of the 64-byte blocks
in all the lanes in the 4 YMM registers and the other second half
in other 4 YMM registers, avoiding a last round of shuffles required
to transpose all the data.

Change-Id: Ib272b6334e6dfbe42609691c667be8959b32135f