Optimize AVX2 4x4 transpose
SHA384 and SHA512 implementations perform an 4x4 64-bit transpose
of the data coming from 4 lanes, before starting to process it.

This transpose is optimized by loading the first half of the 64-byte blocks
in all the lanes in the 2 YMM registers and the other second half
in other 2 YMM registers, avoiding a last round of shuffles required
to transpose all the data.

Change-Id: I2025f16b31ea9cffa878a646c83aa5234f29b719