Return completed job after first job submission
When a job is submitted and is completed straight away,
the job can be returned immediately.
However, the job was not being returned until another job
was submitted, adding latency unnecessarily.

This patch makes it happen and modify the unit tests
to expect jobs to be returned immediately, when the algorithm
permits it (in case of DES, depending on the architecture,
the implementation is single buffer based or multi buffer based,
so the tests need to take into account this fact).

Change-Id: I071ca16fbe6084a73f103987d89d275003c0ed90