Added options to perform safety checks in ZUC
Added compilation options "SAFE_DATA" and "SAFE_PARAM",
used in the ZUC algoritm (will be expanded to other algorithms
in the future).

SAFE_PARAM adds extra checks for input data.
SAFE_DATA clears stack memory and registers,
containing sensitive information, such as IVs or keys,
upon the completion of ZUC routines.

This results in an increase in safety in the code,
with minimal, but potentially noticeable, performance degradation.

Change-Id: I1400d0804d1d4ed60fc79872197500660d7f9d15