Add cross validation application
This new application (ipsec_xvalid_test) performs tests on
various combinations of cipher and hash algorithms,
on different buffer sizes and using different architectures.

The basic test performed is the following:
- The message gets encrypted and a digest is generated from the ciphertext,
  using one architecture
- The output message (ciphertext) gets decrypted and another digest
  is generated from the ciphertext, using another architecture (can be the same
  as the first step)
- The decrypted message is compared against the original plaintext
- Both generated digests are compared
- The test fails if any of these comparison fails

By default, the test loops through:
- All cipher, hash and AEAD algorithms supported
- Buffer sizes between 16 and 2048 bytes, in steps of 16 bytes
- All architectures supported (SSE, AESNI_EMU, AVX, AVX2, AVX512)

Any of the combinations above can be set, performing a single test.
A sample command line to do this is the following:

./ipsec_xvalid_test --job-size 512 --enc-arch AVX512 --dec-arch AVX --aead-algo aes-gcm-128

This command performs a test using AES-GCM, on 512B buffers,
encrypting with AVX512 and decrypting with AVX.

The application can be also launched with no parameters, going through
all cipher+hash combinations and architectures.

Change-Id: I7f5c8e9b66a2312570862ce204fb1ed191f56182