Modify KASUMI API
Instead of using uint8_t * pointers,
use a the more generic void * pointer.
Also, adding const qualifier to those parameters
that should not be changed.
Lastly, the previous implementation was reordering the buffers
passed, modifying the input array of pointers, which is not
a good approach. Instead, pointers are copied internally,
so the input array remains the same.

Change-Id: I44a170d0b11bef3bdd433cf39834e1deeb0d500e