LibTestApp: check KASUMI F8 is not overwritting outside scope
Added extra checks in the test app, verifying that the last
bits are not modified, in case of a partial byte
(when length is non byte multiple) and that full bytes around the buffer
are not modified either.

Test vectors are modified to have the last bits set to 1
(the ones not to be ciphered), as the destination buffer is filled
with ones.

Change-Id: I048a49a1de5a64272831ac55e33287df05299801