LibPerfApp: display progress bar as benchmark runs
Currently the performance application by default runs on
a huge amount of variants, including different algorithms,
key sizes, cipher directions, architectures.
This makes the application looks like it is frozen,
where it is actually benchmarking.

A progress bar displayed as the benchmark runs, to show
some progress to the user for each run.

Change-Id: If682b25bd99f2d39c796c2b711113cc6637f5ff4