ipsec_xvalid: support PON
Added PON algorithm (including no-cipher variant) to the application.
The buffer size passed is the XGEM payload (including CRC if
it is bigger than 4 bytes), and extra bytes are allocated
to include XGEM header and padding (if required).
Therefore, the smallest buffer allocated is 16 bytes long
(8 bytes for header, 1 byte for payload and 7 bytes of padding)

The two variants can be tested either with the default parameters
or with "--aead-algo pon-128" or "--aead-algo pon-128-no-ctr".

Change-Id: I534c150c85111bb9847fe032137f6cc32d48e658