Avoid use of stack in AES-CFB
With the new simd_load_X and simd_store_X macros,
which load and store a variable number of bytes into and from
XMM registers, there is no need to store input buffers
in the stack.
Instead, up to 16 bytes of the input buffer are read directly
into an XMM register and up to 16 bytes are written from
the result, contained in another XMM register.

Change-Id: I0dc6c98c47049e08466aae4887aa98fd9e3c8d25