Fix clearance of CCM sensitive info
In flush CCM job, the IV and initial blocks of
the returned job were not being cleared, since
job_in_lane was non zero.
When clearing the sensitive information of the used
lanes, job_in_lane is checked and only when these
pointers are zero, this data is cleared.
Since this operation was done before the job_in_lane
pointer of the returned job was zeroed, this information
was not clear.

Change-Id: I7cc70ae09241a6a3e1c584d8beb21e7a91e2ea97