ipsec_xvalid: search for sensitive info in MB_MGR
When library is compiled with SAFE_DATA, it is expected
that all sensitive information, such as keys, IVs, etc.
is cleared after a job is returned.

New parameter --safe-check is added in the ipsec_xvalid
application (disabled by default), so when it is set,
a byte pattern is set in keys, plaintext, IVs and the MB_MGR
structure is checked to verify none of this information
remains in it.

Change-Id: I543bbf36f5d4bf7710ef72d68fa9da260ae33649