Fix clearance of DES sensitive info
When flushing DES jobs, the IVs of the NULL lanes
were not being cleared, since the mask register
containing the mask to clear these IVs was being
clobbered in a function call.

Before getting into this function, the mask register
is saved in the stack (through a GP register)
and restored later on before using it to clear
the IVs.

Change-Id: Ie0e1054586e42f81c2b55d6f75cc1e76e3e0b045