SNOW3G: Fix output overwrite in snow3g_1_buffer_bit API
When length of buffer to ciphered is not byte multiple,
the last byte has some bits that needs to be preserved
from output (partial byte).

The code was preserving these last bits from the input
always, and in case of an out-of-place operation,
the bits should be preserved from the output.

Change-Id: I4eac5863c54f2891ce4beabc27b5cc6107113d73