ipsec_xvalid: search for sensitive info in stack
When library is compiled with SAFE_DATA, it is expected
that all sensitive information, such as keys, IVs, etc.
is cleared after a job is returned.

New parameter --safe-check is added in the ipsec_xvalid
application (disabled by default), so when it is set,
a byte pattern is set in keys, plaintext, IVs and the stack
is checked to verify none of this information
remains in it.

Change-Id: I5c4bf1921563277733d92b8346fcf886d4e1348e