ipsec_xvalid: search for sensitive info in registers
When library is compiled with SAFE_DATA, it is expected
that all sensitive information, such as keys, IVs, etc.
is cleared after a job is returned.

New parameter --safe-check is added in the ipsec_xvalid
application (disabled by default), so when it is set,
a byte pattern is set in keys, plaintext, IVs and the
GP and SIMD registers are checked to verify
none of this information remains in it.

Change-Id: Ia8503718fbcde01eb55a5f98e19ae79b79358b58