ipsec_xvalid: clear scratch SIMD registers after setting patterns
The cross validation tool sets patterns in buffers to search
for leftover sensitive data in various memory areas.
These patterns might be stored in SIMD registers, which might
be stored in stack when submitting jobs or expanding keys.
To avoid this problem, scratch SIMD registers are cleared
before calling the Multi Buffer API.

Change-Id: I304c04e21feba48f91c8cf34e4d047e9be17289a