lib: add GHASH API
Add API to perform GHASH on an input buffer,
giving up to 16 bytes of output.

IMB_GHASH() requires gcm_key_data structure to be initialized
calling previously IMB_GCM_PRE_xxx() function, which is the key precompute
function used for AES-GCM, before it is passed as one of the parameters.