Document and use new compilation flag NO_COMPAT_IMB_API_053
There are some public enums that are conflicting
with other enums defined in other libraries, due to
too generic naming.
In order to solve this issue, existing enums will be renamed.

In order to maintain backwards compatilibity for the next few
releases, the previous enums will be maintained,
so existing applications will not be broken.

A new compilation flag, NO_COMPAT_IMB_API_053, is added.
The previous enums can be removed if this compilation flag
is set (which is disabled by default), when building and linking
applications.

For applications which use various libraries, this flag
should be set, so no redefinition of enums will take place, by
passing -DNO_COMPAT_IMB_API_053 to its CFLAGS (or EXTRA_CFLAGS).