test: simplify KASUMI F8 N buffer tests
Allocate, initialize and use a single array of buffers
and iterate through it for all the KASUMI F8 N buffer tests,
instead of using an array of array of buffers.

Thanks to this simplification, some issues reported by Valgrind
due to the complex test (including usage of uninitialized variables,
leading to potential segfaults) are resolved.