test: fix potential buffer overflow in ZUC tests
Maximum number of buffers that can be passed
to the ZUC test functions is MAXBUFS.
The ZUC EIA3 n test checks the number of buffers passed,
and caps it to number of EIA3 test vectors available.
Knowing that this number is higher than MAXBUFS,
this can cause a buffer overflow due to access to
arrays which size is MAXBUFS.

MAXBUFS is calculated then to be maximum of EIA3/EEA3
test vectors, to solve this problem.