avx512: optimize ZUC keystream stores
store4B_kstr16 function is used to store 64B of keystream
for each of the 16 lanes in 16 rounds of 4B each, by extracting
the 4B from a ZMM register containing the keystreams of
all lanes and storing in using the different keystream pointers,
updating also their address.

This method is quite inefficient, as it involves a lot of instructions
and memory loads/stores.

Instead, the following method is used:
- OFS_X3, containing the 4B of keystream for all 16 lanes in each round
  is stored in an ZMM register, which remains untouched in all 16 rounds
- After 16 rounds, all these registers are transposed, so the 64B of
  each lane is in a different ZMM register, which is stored to memory
  with store64B_kstr16
- If number of rounds to do is less than 16, the original store4B_kstr16
  is used.