avx512: optimize LFSR initialization in ZUC
ZUC initialization function sets the LFSR registers
using 15-bit D constants alongside with bytes from the IV
and key of all 16 lanes.

The initialization of these registers was done in an inefficient
way, with a lot of small memory loads and shifts.

Instead, ZMM registers are used to perform this initialization,
for each lane, performing a single read of IV, key and D constant,
and a single 64B store of the LFSR registers for each lane.