avx512: add in-place 16x16 32-bit transpose
Added macro to transpose 16x16 32-bit elements contained
in 16 ZMM registers (TRANSPOSE16_U32). This new macro replaces
the original TRANPOSE16_U32, which has been renamed as
TRANSPOSE16_U32_PRELOADED, since this latter macro requires
the input data to be loaded from memory into the ZMM registers,
where it can be loaded interleaved, avoiding a final round of shuffles.
In case of the new macro, it is assumed that the transpose is done
"in-place", without any loading from memory.