lib: expose some ZUC functions to maximize throughput
Expose internal AVX/AVX2 ZUC-EEA3/EIA3 multi-buffer functions
to the rest of the library, in order to maximize throughput.

For AVX512 implementation, if user calls zuc_eea3_n_buffer with
a non-multiple of 16 number of buffers, only multiples of 16 buffers
will be handled by the multi-buffer implementation, and the rest
will be handled one by one.

Instead of this, batches of 8 and 4 buffers can be handled
using the AVX2 and AVX multi-buffer implementations.

This approach is used also for EIA3 and in AVX2.