lib: fix block count increment in AES-CTR (fix for issue #43)
AES-CTR algorithm used in IPSec (cipher_mode = AES_CNTR) uses a counter block with 4 bytes for block count,
according to the following RFC: https://tools.ietf.org/html/rfc3686#section-2.1

AES-CTR algorithm used in PDCP (AES_CNTR_BITLEN) uses a block count with 8 bytes,
according to section B.1.3 of the following document:
https://www.etsi.org/deliver/etsi_ts/133400_133499/133401/14.05.00_60/ts_133401v140500p.pdf

The latter algorithm was added in release 0.53, and introduced issues in both AES_CNTR and AES_CNTR_BITLEN.

The issue is in the way this block count is being incremented in all implementations.
For AES_CNTR, it should increment only 4 bytes, and for AES_CNTR_BITLEN, 8 bytes.
In case of AVX, it was actually doing it the other way, incrementing 8 bytes
for AES_CNTR and 4 bytes for AES_CNTR_BITLEN.
In case of the other implementations, SSE and AVX512-VAES, AES_CNTR was correct,
but AES_CNTR_BITLEN was not, since it was incrementing only 4 bytes.