lib: make OOO managers private
Out-of-order (OOO) managers do not need to be
exposed publicly, as they should be used only internally
in the library.
Therefore, the OOO manager definitions are moved to
an internal header file and pointers to these are left
in IMB_MGR structure.

alloc_mb_mgr() function allocates then memory for the structure
and the OOO managers.

free_mb_mgr() function frees the memory used for the structure
and the OOO managers.