lib: support stand-alone AES-GMAC
Added enums to perform AES-GMAC without needing
to be coupled to GCM. Key size is selected using
the different enums:
- IMB_AUTH_AES_GMAC_128
- IMB_AUTH_AES_GMAC_192
- IMB_AUTH_AES_GMAC_256

Inside the job structure, key, IV and IV size can
be passed through the fields in "u.GMAC".

This allows the algorithm to be used with other cipher
algorithm, as the fields enc_keys, dec_keys, iv and iv_len_in_bytes
are free to be used for the cipher algorithm.