avx2: optimize LFSR initialization in ZUC
ZUC initialization function sets the LFSR registers
using 15-bit D constants alongside with bytes from the IV
and key of all 8 lanes.

The initialization of these registers was done in an inefficient
way, with a lot of small memory loads and shifts.

Instead, YMM registers are used to perform this initialization,
for each lane, performing two reads of IV, key and D constant,
and two 32B store of the LFSR registers for each lane.