avx2: accept any buffer size in ZUC encryption for job API
Instead of accepting only multiple of 4 bytes, accept any buffer size
which is the common length for all the lanes.
When the length is not multiple of 4 bytes, the length to encrypt
is rounded up to the nearest size multiple of 4 bytes and buffers
which length is at least that size are encrypted up to that size.
For the other buffers, only their size will be encrypted.