test: fix FreeBSD compilation
Use sighandler_t only on Linux, as it is the only OS
that has this type (use void * if other OS).

Also, this commit fixes a warning on function "seg_handler",
in direct_api.test.c.
function 'seg_handler' could be declared with attribute 'noreturn' [-Wmissing-noreturn]

This commit is part of fix for #62.