avx512: initialize ZUC state in ZUC-EIA3 submit/flush
Separate ZUC initialization from the authentication function,
so it can be called from job submit/flush function.
This way, only the state for new buffers will be initialized.

This commit also simplifies the flush logic in EEA3 and fixes
an issue when using bsf instruction to get the first bit set.