avx512: generate ZUC keystream on specified lanes
Generate 8 bytes of ZUC keystream for specified lanes,
passing a bitmask selecting such lanes. This will be useful
in next commits, where some lanes will use saved bytes from
previous calls and new lanes will require new 8 bytes.

The function actually outputs keystreams in all lanes,
but the states for the disabled lanes are not modified,
meaning that the keystream of the disabled lanes is garbage
and the keystream generation may be resumed in next calls.