avx512: perform only multi-buffer ops in ZUC-EIA3 in job API
Modified the crypto code for ZUC-EIA3 in job API to perform
authentication on the common length of all the buffer availables.
For the buffers which data has been authenticated fully, the final
digest is calculated and output in submit/flush function.
For the rest of the buffers, their final 8 bytes of keystreams
are not used and they are saved to be used in next calls, where
their authentication is resumed (once more new buffers are submitted).