lib: fix AES-CBCS in SSE/AVX
When passing multiple buffers with different buffer sizes,
the common length of them would be processed, as in CBC mode.
The problem is that the input/output pointers will be moved
equally in all buffers, meaning that the pointer of
a non-completed buffer might be pointing at a block that should
not be encrypted, and therefore, it will be encrypted in the next call.

Instead of this, the pointers are moved in steps of 160 bytes,
making sure that for non-new buffers, the input/output pointer
points at the next block to be encrypted