test: [API] add reset API test
Add tests to validate reset API (imb_set_pointers_mb_mgr function).
First, this function is called to reset the function pointers,
but without resetting OOO managers.
Then, every time a job is submitted and it is not returned straight
away, it is expected that such job is still in an OOO manager,
so the job should be returned after flushing, even when the
reset API is called (as long as the option to resetting the whole
IMB_MGR is disabled).