lib/mb_mgr_burst_async: cover full queue still processing case
It is possible for the MB_MGR job queue to have all its jobs busy,
so `submit_burst_and_check()` will not be able to return number
of completions greater than 0. In that case, instead of marking
the queue as "empty", just return 0 and check if we ran into
this corner case during job flush.

Signed-off-by: Krzysztof Karas <krzysztof.karas@intel.com>