vaes-avx512: [gcm] enable small packet optimization for all available single call API variants
The optimization is enabled for:
- single call cases
- any IV size (variable IV interface) and 12 byte IV size
- both cipher directions

To save space small packet code generation is disabled in GCM_ENC_DEC in in the above cases.
This is no longer needed as new small packet code covers it.

GCM context got removed from argument list for GCM_ENC_DEC_0_TO_256 macro (not required).