test: [xvalid-app] add offset (optional) to source buffers
Add offset to source/destination buffers, to test if the offset parameters
in the job structure are taken into account.

The application adds the offset directly to job->src and job->dst in one cipher
direction and sets the job->cipher_start_src_offset_in_bytes and/or
job->hash_start_src_offset_in_bytes in the other direction.

This offset is 4 bytes by default, and it can be changed with
the --offset parameter from the command line.

E.g.: ./imb-xvalid --aead-algo aes-gcm-128 --offset 16

Signed-off-by: Pablo de Lara <pablo.de.lara.guarch@intel.com>